extracted ExportPlotModal into its own component
